+++ packages/editor-core/src/editor/plugins/insert-block/index.tsx
@@ -6,10 +6,11 @@ import { stateKey as tablesStateKey } from '../../../plugins/table';
 import ToolbarInsertBlock from '../../../ui/ToolbarInsertBlock';
 
 const insertBlockPlugin: EditorPlugin = {
-  primaryToolbarComponent(editorView) {
+  primaryToolbarComponent(editorView, eventDispatcher, providerFactory, appearance) {
+    const isCommentAppearance = appearance === 'comment';
     const pluginStateBlockType = blockTypeStateKey.getState(editorView.state);
     const pluginStateTable = tablesStateKey.getState(editorView.state);
-    const pluginStateMedia = mediaStateKey.getState(editorView.state);
+    const pluginStateMedia = !isCommentAppearance ? mediaStateKey.getState(editorView.state) : undefined;
     return <ToolbarInsertBlock
       editorView={editorView}
       pluginStateBlockType={pluginStateBlockType}