+++ packages/editor-core/src/editor/types/index.ts
@@ -3,3 +3,4 @@ export * from './editor-config';
 export * from './editor-plugin';
 export * from './editor-props';
 export * from './editor-appearance-component';
+export * from './macros';