+++ packages/editor-core/stories/00-comment-editor-story.tsx
@@ -1,0 +1,86 @@
+import { action, storiesOf } from '@kadira/storybook';
+import * as React from 'react';
+
+import Button, { ButtonGroup } from '@atlaskit/button';
+import Editor from './../src/editor';
+import EditorContext from './../src/editor/ui/EditorContext';
+import WithEditorActions from './../src/editor/ui/WithEditorActions';
+import ToolsDrawer from './ToolsDrawer';
+import { name, version } from '../package.json';
+import { storyDecorator } from '../src/test-helper';
+import CollapsedEditor from '../src/editor/ui/CollapsedEditor';
+
+const SAVE_ACTION = () => action('Save')();
+const CANCEL_ACTION = () => action('Cancel')();
+const EXPAND_ACTION = () => action('Expand')();
+const analyticsHandler = (actionName, props) => action(actionName)(props);
+const exampleDocument = {
+  version: 1,
+  type: 'doc',
+  content: [
+    {
+      type: 'paragraph',
+      content: [
+        { type: 'text', text: 'Some example document with emojis ' },
+        {
+          type: 'emoji',
+          attrs: { shortName: ':catchemall:', id: 'atlassian-catchemall', text: ':catchemall:' }
+        },
+        { type: 'text', text: ' and mentions ' },
+        {
+          type: 'mention',
+          attrs: { id: '0', text: '@Carolyn', accessLevel: '' }
+        },
+        { type: 'text', text: '. ' }
+      ]
+    }
+  ]
+};
+
+storiesOf(name, module)
+  .addDecorator(storyDecorator(version))
+  .add('Comment Editor', () =>
+    <EditorContext>
+      <div>
+        <WithEditorActions
+          // tslint:disable-next-line:jsx-no-lambda
+          render={actions =>
+            <ButtonGroup>
+              <Button onClick={() => actions.replaceDocument(exampleDocument)}>Load Document</Button>
+              <Button onClick={() => actions.clear()}>Clear</Button>
+            </ButtonGroup>
+          }
+        />
+        <ToolsDrawer
+          // tslint:disable-next-line:jsx-no-lambda
+          renderEditor={({ mentionProvider, emojiProvider, mediaProvider, onChange }) =>
+            <div style={{ padding: '20px' }}>
+              <CollapsedEditor
+                placeholder="What do you want to say?"
+              >
+                <Editor
+                  appearance="comment"
+                  analyticsHandler={analyticsHandler}
+                  shouldFocus={true}
+
+                  allowTextFormatting={true}
+                  allowTasksAndDecisions={true}
+                  allowHyperlinks={true}
+                  allowCodeBlocks={true}
+                  allowLists={true}
+
+                  mentionProvider={mentionProvider}
+                  emojiProvider={emojiProvider}
+                  mediaProvider={mediaProvider}
+
+                  onChange={onChange}
+                  onSave={SAVE_ACTION}
+                  onCancel={CANCEL_ACTION}
+                  onExpand={EXPAND_ACTION}
+                />
+              </CollapsedEditor>
+            </div>}
+        />
+      </div>
+    </EditorContext>
+  );